feat(keyboard): update corne qmk layout
Replace dvorak layer with 2nd YulianKb layer with 2 ctrl keys on the
home row for each hand.  This is an experiment to try and make switching
between mac and linux more tolerable.  The first layer will be for mac
and the second for linux.

This also swaps 0 and = on the number layer.